+++ testprogs/opentest.c
@@ -59,7 +59,7 @@ main(int argc, char **argv)
 {
 	register int op;
 	register char *cp, *device;
-	int dorfmon, dopromisc, snaplen, useactivate, bufsize;
+	int dorfmon, dopromisc, snaplen, useactivate, useopen, bufsize;
 	char ebuf[PCAP_ERRBUF_SIZE];
 	pcap_if_t *devlist;
 	pcap_t *pd;
@@ -71,13 +71,14 @@ main(int argc, char **argv)
 	snaplen = MAXIMUM_SNAPLEN;
 	bufsize = 0;
 	useactivate = 0;
+	useopen = 0;
 	if ((cp = strrchr(argv[0], '/')) != NULL)
 		program_name = cp + 1;
 	else
 		program_name = argv[0];
 
 	opterr = 0;
-	while ((op = getopt(argc, argv, "i:Ips:aB:")) != -1) {
+	while ((op = getopt(argc, argv, "i:Iops:aB:")) != -1) {
 		switch (op) {
 
 		case 'i':
@@ -87,6 +88,10 @@ main(int argc, char **argv)
 		case 'I':
 			dorfmon = 1;
 			useactivate = 1;	/* required for rfmon */
+			break;
+
+		case 'o':
+			useopen = 1;
 			break;
 
 		case 'p':
@@ -181,6 +186,16 @@ main(int argc, char **argv)
 			warning("%s: %s\n(%s)", device,
 			    pcap_statustostr(status), pcap_geterr(pd));
 		} else
+			printf("%s opened successfully\n", device);
+	} else if (useopen) {
+		*ebuf = '\0';
+		pd = pcap_open(device, 65535, PCAP_OPENFLAG_PROMISCUOUS,
+		    1000, NULL, ebuf);
+		if (pd == NULL)
+			error("%s", ebuf);
+		else if (*ebuf)
+			warning("%s", ebuf);
+		else
 			printf("%s opened successfully\n", device);
 	} else {
 		*ebuf = '\0';
